add register packet, etc roomService enhancement

+++ src/service/SocketService.ts
@@ -3,37 +3,73 @@ import { v4 as uuidv4 } from "uuid";
 import { Game } from "./GameService";
 import { DatabaseService } from "./DatabaseService";
 import { Room, RoomService } from "./RoomService";
+import { Socket } from "socket.io";
+
+interface CustomSocket extends Socket {
+    userId?: number; //
+}
+
 export namespace SocketService {
     export async function initialize() {
         listen();
     }
     function listen() {
-        Vars.io.on("connection", (socket) => {
-            socket.on("LOGIN", async ({ userName, phoneNumber }: LoginPacket, callback) => {
-                const { userId } = await DatabaseService.createUser({ userName, phoneNumber });
-                console.log(userId);
-                RoomService.users.push({
-                    userId,
-                    userName,
-                    phoneNumber,
-                    score: 0,
-                    teamId: 0,
-                    teamName: "",
-                });
-                callback({ userId: userId });
+        Vars.io.on("connection", (socket: CustomSocket) => {
+            socket.on("disconnect", async (reason) => {
+                console.log("socket disconnected : ", reason);
+
+                if (!socket.userId) return;
+
+                if (socket.rooms.has("lobby")) {
+                    Vars.io.emit("LEAVELOBBY", socket.userId);
+                }
+
+                const room = RoomService.getRoomFromUserId(socket.userId);
+                if (room) {
+                    await RoomService.deleteRoom(room.roomId);
+                    socket.broadcast.to("lobby").emit("DELETEROOM", { roomId: room.roomId });
+                    socket.broadcast.to(room.roomId.toString()).emit("DELETEROOM", {});
+                }
             });
-            socket.on("DELETEROOM", async ({ roomId, userId }: DeleteRoomPacket, callback) => {
+
+            socket.on("LOGIN", async ({ userName, phoneNumber }: LoginPacket, callback) => {
+                const result = await DatabaseService.findUser({ userName, phoneNumber });
+                if (result.success) {
+                    RoomService.addUser(result.user!);
+                    socket.userId = result.user?.userId;
+                    await socket.join("lobby");
+                    Vars.io.to("lobby").emit("JOINLOBBY", result.user);
+                }
+                callback(result);
+            });
+
+            socket.on("REGISTER", async ({ userName, phoneNumber }: LoginPacket, callback) => {
+                const result = await DatabaseService.createUser({ userName, phoneNumber });
+                if (result.success) {
+                    RoomService.users.push({
+                        userId: result.userId,
+                        userName,
+                        phoneNumber,
+                        score: 0,
+                    });
+                }
+                callback(result);
+            });
+
+            socket.on("DELETEROOM", async ({ roomId }: DeleteRoomPacket, callback) => {
                 const room = RoomService.getRoom(roomId);
-                const user = room?.getUser(userId);
+                if (!socket.userId) return;
+                const user = room?.getUser(socket.userId);
                 if (!user || user.power !== "leader") {
                     callback({ success: false });
                     return;
                 }
                 await RoomService.deleteRoom(roomId);
+                socket.broadcast.to("lobby").emit("DELETEROOM", { roomId });
                 socket.broadcast.to(roomId.toString()).emit("DELETEROOM", {});
                 callback({ success: true });
             });
-            /**                          방아이디, 강퇴하는 유저아이디, 강퇴당하는 유저아이디 */
+
             socket.on("KICKMEMBER", async ({ roomId, userId, targetId }, callback) => {
                 const room = RoomService.getRoom(roomId);
                 if (!room) {
@@ -48,16 +84,10 @@ export namespace SocketService {
                 room.kickMember(targetId);
                 callback({ success: true });
             });
-            socket.on("GETROOMS", async ({ userId }: GetRoomsPacket, callback) => {
-                const user = RoomService.getUser(userId);
-                if (!user) {
-                    callback({
-                        success: false,
-                    });
-                    return;
-                }
+
+            socket.on("GETROOMS", async ({}: GetRoomsPacket, callback) => {
+                if (!socket.userId) return;
                 const rooms = RoomService.getRooms();
-                console.log(rooms);
                 callback({
                     success: true,
                     rooms: rooms.map((room) => ({
@@ -68,16 +98,18 @@ export namespace SocketService {
                     })),
                 });
             });
-            socket.on("STARTGAME", async ({ userId, roomId }: StartGamePacket, callback) => {
+
+            socket.on("STARTGAME", async ({ roomId }: StartGamePacket, callback) => {
                 const room = RoomService.getRoom(roomId);
                 if (!room) return;
-                const user = room.getUser(userId);
+                const user = room.getUser(socket.userId);
                 if (!user || user.power !== "leader") return;
                 await room.getGame().startGame();
                 callback({ success: true });
             });
-            socket.on("JOINROOM", async ({ userId, roomId }: JoinRoomPacket, callback) => {
-                const user = RoomService.getUser(userId);
+
+            socket.on("JOINROOM", async ({ roomId }: JoinRoomPacket, callback) => {
+                const user = RoomService.getUser(socket.userId);
                 if (!user) {
                     callback({ success: false });
                     return;
@@ -87,7 +119,7 @@ export namespace SocketService {
                     callback({ success: false });
                     return;
                 }
-                const success = RoomService.joinRoom(user, roomId);
+                const success = RoomService.joinRoom(user.userId, roomId);
                 if (!success) {
                     callback({ success: false });
                     return;
@@ -95,10 +127,13 @@ export namespace SocketService {
                 const users = room.getUsers();
                 callback({ success: true, users: users });
                 socket.broadcast.to(roomId.toString()).emit("JOINUSER", user);
+                socket.leave("lobby");
                 socket.join(roomId.toString());
+                socket.broadcast.to("lobby").emit("JOINUSER", { user, roomId });
             });
-            socket.on("CREATEROOM", async ({ userId, roomName }: CreateRoomPacket, callback) => {
-                const user = RoomService.getUser(userId);
+
+            socket.on("CREATEROOM", async ({ roomName }: CreateRoomPacket, callback) => {
+                const user = RoomService.getUser(socket.userId);
                 if (!user) {
                     callback({ success: false });
                     return;
@@ -106,12 +141,24 @@ export namespace SocketService {
                 const roomId = uuidv4();
                 const game = new Game(roomId);
                 game.addUser(user);
+
                 const room = new Room(game, roomId, roomName);
                 RoomService.addRoom(room);
                 room.addUser(user, "leader");
-                callback({ room: room });
+
+                socket.leave("lobby");
+                callback({ room });
+
+                Vars.io.to("lobby").emit("CREATEROOM", {
+                    roomId: room.roomId,
+                    roomName: room.roomName,
+                    users: room.getUsers(),
+                    started: room.getGame().isStarted,
+                });
             });
-            socket.on("CHAT", ({ userInfo, roomId, message }: ChatPacket, callback) => {
+
+            socket.on("CHAT", ({ roomId, message }: ChatPacket, callback) => {
+                const user = RoomService.getUser(socket.userId);
                 const room = RoomService.getRoom(roomId);
                 if (!room) {
                     callback({
@@ -119,15 +166,20 @@ export namespace SocketService {
                     });
                     return;
                 }
-                socket.broadcast.to(roomId).emit("CHAT", { userId: userInfo.userId, message });
+                if (!user) {
+                    callback({ success: false });
+                    return;
+                }
+                socket.broadcast.to(roomId).emit("CHAT", { userName: user.userName, message });
             });
-            socket.on("WORD", ({ userId, roomId, word }: WordPacket, callback) => {
+
+            socket.on("WORD", ({ roomId, word }: WordPacket, callback) => {
                 const room = RoomService.getRoom(roomId);
                 if (!room) return;
-                const user = room.getUser(userId);
+                const user = room.getUser(socket.userId);
                 if (!user) return;
-                const success = room.getGame().word(userId, word);
-                socket.broadcast.to(room.roomId).emit("WORD", { userId, success, word });
+                const success = room.getGame().word(socket.userId!, word);
+                socket.broadcast.to(room.roomId).emit("WORD", { userId: socket.userId!, success, word });
             });
         });
     }
